Split Helpers::Controller into submodules, drop 3 Metrics overrides
Helpers::Controller had grown to a 294-line module mixing four
concerns, held together by Metrics/ModuleLength Max: 209 in
.rubocop_todo.yml, plus CyclomaticComplexity Max: 13 and
PerceivedComplexity Max: 10 for its two biggest methods.

Split it by concern, all included into the prepended Controller module
(included modules sit between the prepended module and Doorkeeper's
Helpers::Controller in the ancestor chain, so super still reaches the
Doorkeeper implementation):

- Prompt        - prompt= handling (none/login/consent/select_account),
                  prompt value parsing, reauthentication and account
                  selection round-trips
- MaxAge        - max_age= staleness enforcement and auth_time
                  resolution (auth_time_from_session with legacy
                  fallback)
- ErrorResponse - rendering OIDC authorization errors through
                  Doorkeeper's OAuth response objects
- TokenMatching - subset-scope token matching for prompt=none / consent
                  skipping (issue #63) and the skip_authorization?
                  override

The two complexity hotspots are decomposed: the prompt case dispatch
moved into apply_oidc_prompt! with per-value handlers, and
handle_oidc_max_age_param! now delegates to normalized_oidc_auth_time /
oidc_auth_time_stale?. handle_oidc_error! reuses clear_oidc_response
instead of duplicating it.

Metrics/ModuleLength, Metrics/CyclomaticComplexity and
Metrics/PerceivedComplexity now pass at RuboCop defaults and their
overrides are removed from .rubocop_todo.yml. MethodLength Max: 33 and
AbcSize Max: 35 remain, accommodating DiscoveryController's
provider_response hash (same accommodation upstream doorkeeper makes
for its own config-shaped methods).

Behavior-preserving: 303 examples green on latest doorkeeper and on
doorkeeper 5.8.0, rubocop clean.
